+++ android/src/main/java/com/hoxfon/react/RNTwilioVoice/TwilioVoicePackage.java
@@ -12,6 +12,14 @@ import java.util.List;
 
 public class TwilioVoicePackage implements ReactPackage {
 
+    private boolean mShouldAskForPermission;
+    public TwilioVoicePackage() {
+        mShouldAskForPermission = true;
+    }
+
+    public TwilioVoicePackage(boolean shouldAskForPermissions) {
+        mShouldAskForPermission = shouldAskForPermissions;
+    }
     // Deprecated in RN 0.47.0
     public List<Class<? extends JavaScriptModule>> createJSModules() {
         return Collections.emptyList();
@@ -25,8 +33,7 @@ public class TwilioVoicePackage implements ReactPackage {
     @Override
     public List<NativeModule> createNativeModules(ReactApplicationContext reactContext) {
         List<NativeModule> modules = new ArrayList<>();
-        modules.add(new TwilioVoiceModule(reactContext));
+        modules.add(new TwilioVoiceModule(reactContext, mShouldAskForPermission));
         return modules;
     }
-
 }