+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 [Release Section](https://github.com/hoxfon/react-native-twilio-programmable-voice/releases)
 
+## 3.21.0
+- Android: allow to pass arbitrary parameters to call voice.call() as it is on iOS
+
+## 3.20.1
+- iOS: fix crash when callSid is nil for CallInviteCanceled
+
+## 3.20.0
+- Android: option to opt out microphone permission request
+
+## 3.19.0
+- upgrade com.twilio:voice-android to 2.0.7
+- upgrade firebase-messaging to 17.+
+
 ## 3.18.1
 - Validate token type before calling native module